requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.6
aiohttp-socks==0.8.3
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
annoy==1.17.3
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
asteval==0.9.31
astropy==5.0
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.0.0
blinker==1.6.3
bokeh==3.3.2
boto3==1.28.85
botocore==1.31.85
Brlapi==0.8.3
Brotli==1.0.9
CacheControl==0.12.10
cachy==0.3.0
catboost==1.2.2
category-encoders==2.6.3
ccsm==0.9.14.1
certifi==2023.11.17
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
cleo==0.8.1
click==8.1.7
clikit==0.6.2
cloudpickle==2.2.1
colorama==0.4.4
colorcet==3.0.1
colorlog==6.8.0
comm==0.1.4
command-not-found==0.3
compizconfig-python==0.9.14.1
contourpy==1.2.0
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
cupy-cuda11x==12.3.0
cycler==0.12.1
daal==2024.0.1
daal4py==2024.0.1
databricks-cli==0.18.0
DAWG-Python==0.7.2
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dp==1.0
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fastapi==0.104.1
fasteners==0.14.1
fastjsonschema==2.18.0
fastrlock==0.8.2
filelock==3.6.0
flake8==6.1.0
Flask==2.3.3
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.12
fsspec==2023.12.2
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.40
glob2==0.7
graphviz==0.20.1
greenlet==3.0.0
gunicorn==21.2.0
h11==0.13.0
holoviews==1.18.1
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.20.2
hvplot==0.9.0
idna==3.6
ImageHash==4.3.1
imageio==2.33.1
importlib-metadata==4.6.4
intervaltree==3.1.0
ipykernel==6.25.1
ipymarkup==0.9.0
ipython==8.15.0
ipython_genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.24.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lightgbm==4.2.0
lime==0.2.0.1
linkify-it-py==2.0.2
llvmlite==0.41.1
lmfit==1.0.3
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mglearn==0.2.0
missingno==0.5.2
mistune==3.0.1
mlflow==2.7.1
mlxtend==0.23.0
moddb==0.8.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
natasha==1.6.0
navec==0.10.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
notebook==7.0.3
notebook_shim==0.2.3
numba==0.58.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
openshot-qt==3.1.1
optuna==3.5.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
panel==1.3.4
param==2.0.1
paramiko==2.9.3
parso==0.8.3
pastel==0.2.1
patsy==0.5.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==3.10.0
plotly==5.18.0
ply==3.11
poetry==1.1.12
poetry-core==1.0.7
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.12.4
pseudoslit==0.0.2
psutil==5.9.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyarrow==13.0.0
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.21
pyct==0.5.0
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pyerfa==2.0.0.3
pyflakes==3.1.0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==1.26.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyrate-limiter==2.8.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.24
python-socks==2.4.4
pythran==0.10.0
pytz==2023.3.post1
pyviz_comms==3.0.0
PyWavelets==1.5.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
querystring-parser==1.2.4
razdel==0.5.0
referencing==0.30.2
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
s3transfer==0.7.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.1
scikit-learn-intelex==2024.0.1
scipy==1.11.4
scour==0.38.2
screen-resolution-extra==0.0.0
seaborn==0.13.1
SecretStorage==3.3.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.4.3
setproctitle==1.2.2
shap==0.44.0
shellingham==1.4.0
six==1.16.0
sla==1.4.0
slicer==0.0.7
slovnet==0.6.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
sympy==1.9
systemd-python==234
tabulate==0.9.0
tbb==2021.11.0
tenacity==8.2.3
terminado==0.17.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.9.2
toolz==0.11.2
torch==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.62.3
traitlets==5.9.0
transformers==4.36.2
triton==2.1.0
typing_extensions==4.7.1
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
uncertainties==3.1.7
unicodedata2==14.0.0
unity-tweak-tool==0.0.7
uri-template==1.3.0
urllib3==2.0.7
usb-creator==0.3.7
uvicorn==0.15.0
virtualenv==20.13.0+ds
vorbin==3.1.4
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wsproto==1.0.0
xdg==5
xgboost==2.0.3
xkit==0.0.0
xlrd==1.2.0
xyzservices==2023.10.1
yargy==0.16.0
yarl==1.9.4
zipp==1.0.0
